fix(sdk-go): Change `int` args to match proto (#1264)

+++ sdk-go/littlehorse/wf_lib_internal.go
@@ -1054,7 +1054,7 @@ func (t *WorkflowThread) overrideTaskExponentialBackoffPolicy(taskNodeOutput *Ta
 	node.GetTask().ExponentialBackoff = policy
 }
 
-func (t *WorkflowThread) addTimeoutToExtEvt(nodeOutput *NodeOutput, timeoutSeconds int) {
+func (t *WorkflowThread) addTimeoutToExtEvt(nodeOutput *NodeOutput, timeoutSeconds int64) {
 	t.checkIfIsActive()
 
 	node := t.spec.Nodes[nodeOutput.nodeName]
